+++ RELEASE_NOTES.md
@@ -1,5 +1,10 @@
 # Change log
 
+## in develop
+
+* Allows automatic coersion of `read_*` return values from `String` to other primitive types
+* Fixes handling of relative imports that are not in a subdirectory of the main WDL's parent directory
+
 ## 0.14.0 (2021-05-27)
 
 * Adds `fix` tool to automatically fix some non-compliant WDL syntax